Data extractor type class fix
